backporting: Clean tmp files after backport with conflicts
During the backport process, when cherry-picking commits to backport,
the cherry-pick script may fail if there are conflicts. In that case,
the temporary file holding the backported commit is not clean up.

This commit fixes it to clean the temporary file even in case of
failure, while preserving the error code from 'git am'.

Signed-off-by: Paul Chaignon <paul@cilium.io>